requirements.txt
scikit-bio==0.5.8
python-snappy==0.6.1
torch==1.12.0
torchtext==0.13.0

# conda install cudatoolkit=11.6 -c pytorch
# pip install torch==1.12.0+cu116 --extra-index-url https://download.pytorch.org/whl/cu116